pyproject.toml
[project]
authors = [{ name = "McCoy Becker", email = "mccoyb@mit.edu" }]
name = "genjax"
requires-python = ">= 3.12"
version = "0.1.0"
dependencies = [ "penzai>=0.2.5,<0.3", "beartype>=0.20.2,<0.21", "jaxtyping>=0.3.1,<0.4", "tensorflow-probability>=0.25.0,<0.26", "genstudio>=2025.3.12,<2026",
]

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["osx-arm64"]

[tool.pixi.pypi-dependencies]
genjax = { path = ".", editable = true }

[tool.pixi.dependencies]
jax = ">=0.5.2,<0.6"

[tool.vulture]
make_whitelist = true
min_confidence = 80
paths = ["src"]
sort_by_size = true

[tool.pixi.feature.format.tasks]
format = "ruff format . && ruff check . --fix"
vulture = "vulture"

[tool.pixi.feature.docs.tasks]
deploy = "quarto publish gh-pages ./quarto"
preview = "quarto preview ./quarto"

[tool.pixi.environments]
format = { features = ["format"], solve-group = "default" }
docs = { features = ["docs"], solve-group = "docs" }

[dependency-groups]
format = ["ruff>=0.9.2,<0.10","vulture>=2.14,<3"]
docs = ["jupyter"]
